+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.6.0] - 2022-10-07
+
+- [DemoApp] #250 Add bottom sheet to allow the user to show/hide the various options on CardTitleFirst 
+- [DemoApp] #251 Add bottom sheet to allow the user to show/hide the various options on CardImageFirst
+- [DemoApp] #249 Add bottom sheet to allow the user to show/hide the subtitle in SmallCard Variant Page 
+- [DemoApp/SDK] #222 Update the Button page to list the available button types
+- [SDK] Bug #96 Add Small Card 
+- [DemoApp] #248 Technical code refactoring for component page presentation 
+- [SDK/DemoApp] #229 Technical update to use Small and ImageFirst cards in demo app and modules 
+- [DemoApp] #228 Change Chip to switch on List screen for "SecondLineOfText" and "LeadingImage" options 
+- [SDK/DemoApp] #237 Define a Grid of small cards module and use it in application for components
+- [SDK/DemoApp] #182 Add card title first
+- [SDK/DemoApp] #237 Define a Grid of small cards module and use it in application for components 
+- [DemoApp] #235 Rename Module called Card View to Card Collections (2)
+- [SDK] Bug #96 Add Small Card
+
 ## [0.5.0] - 2022-09-02
 
 - [DemoApp] Bug #219 Rename Module called Card View to Card Collections